+++ src/kv/kv_types.h
@@ -702,8 +702,7 @@ namespace ccf::kv
       virtual ~AbstractSnapshot() = default;
       [[nodiscard]] virtual Version get_version() const = 0;
       virtual std::vector<uint8_t> serialise(
-        const std::shared_ptr<AbstractTxEncryptor>& encryptor,
-        size_t max_transaction_size) = 0;
+        const std::shared_ptr<AbstractTxEncryptor>& encryptor) = 0;
     };
 
     virtual ~AbstractStore() = default;